Adds content to feed section, based on fetched posts
